+++ src/components/theme-switch/index.ts
@@ -1,0 +1,1 @@
+export { default as VfThemeSwitch } from "./VfThemeSwitch.vue";